Feature fe disign#2

See merge request sw_track/class_04/web_project/team18/gonggan_mazzip!75

+++ src/views/cart/cart.js
@@ -17,13 +17,13 @@ function calculateTotal() {
 
   checkboxes.forEach((checkbox, index) => {
     if (checkbox.checked) {
-      const item = cartItems[index]; // 이 부분을 수정했습니다.
+      const item = cartItems[index];
       total += item.price * item.quantity;
     }
   });
 
-  totalEl.textContent = `${total.toLocaleString()}원`;
-  orderEl.textContent = `${total.toLocaleString()}원`;
+  totalEl.textContent = `${total.toLocaleString()} 원`;
+  orderEl.textContent = `${total.toLocaleString()} 원`;
 }
 
 const productHtml = cartItems
@@ -39,7 +39,7 @@ const productHtml = cartItems
           <div class="product-detail">
             <div class="product-company">${item.company}</div>
             <div class="product-name">${item.productName}</div>
-            <div class="product-price">${item.price.toLocaleString()}</div>
+            <div class="product-price">${item.price.toLocaleString()} 원</div>
             <div class="product-quantity">
               <button type="button" class="minusBtn">-</button>
               <span class="quantity">${item.quantity}</span>
@@ -143,19 +143,22 @@ if (cartItems.length === 0) {
     const closeButton = item.querySelector('.close-button');
 
     closeButton.addEventListener('click', () => {
-      // 상품 삭제
-      productList.removeChild(item);
-      cartItems.splice(index, 1);
+      const confirmed = confirm('상품을 장바구니에서 삭제 하시겠어요?');
+      if (confirmed) {
+        // 상품 삭제
+        productList.removeChild(item);
+        cartItems.splice(index, 1);
 
-      // 로컬 스토리지 업데이트
-      localStorage.setItem('cartItems', JSON.stringify(cartItems));
+        // 로컬 스토리지 업데이트
+        localStorage.setItem('cartItems', JSON.stringify(cartItems));
 
-      // 총액 계산
-      calculateTotal();
+        // 총액 계산
+        calculateTotal();
 
-      // 로컬 스토리지가 비었을 경우 showEmptyCartMessage 함수 실행
-      if (cartItems.length === 0) {
-        showEmptyCartMessage();
+        // 로컬 스토리지가 비었을 경우 showEmptyCartMessage 함수 실행
+        if (cartItems.length === 0) {
+          showEmptyCartMessage();
+        }
       }
     });
   });
